HSEARCH-3605 Restructure packages for builtin bridge annotations and programmatic APIs (GeoPointBridge)
Let's make sure that, if one day we have more than one builtin bridge,
we'll be able to put all the APIs in the same package so that they're
more easily found.
