+++ mapper/pojo/src/main/java/org/hibernate/search/mapper/pojo/bridge/builtin/annotation/GeoPointBridge.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.bridge.builtin.spatial.annotation;
+package org.hibernate.search.mapper.pojo.bridge.builtin.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
