+++ mapper/pojo/src/main/java/org/hibernate/search/mapper/pojo/bridge/builtin/spatial/impl/LatitudeMarker.java
@@ -6,9 +6,9 @@
  */
 package org.hibernate.search.mapper.pojo.bridge.builtin.spatial.impl;
 
-import org.hibernate.search.mapper.pojo.bridge.builtin.spatial.GeoPointBridgeBuilder;
-import org.hibernate.search.mapper.pojo.bridge.builtin.spatial.LatitudeLongitudeMarkerBuilder;
-import org.hibernate.search.mapper.pojo.bridge.builtin.spatial.annotation.Latitude;
+import org.hibernate.search.mapper.pojo.bridge.builtin.programmatic.GeoPointBridgeBuilder;
+import org.hibernate.search.mapper.pojo.bridge.builtin.programmatic.LatitudeLongitudeMarkerBuilder;
+import org.hibernate.search.mapper.pojo.bridge.builtin.annotation.Latitude;
 import org.hibernate.search.mapper.pojo.bridge.mapping.programmatic.AnnotationMarkerBuilder;
 import org.hibernate.search.mapper.pojo.bridge.mapping.programmatic.MarkerBuildContext;
